types

+++ list.topic.php
@@ -39,7 +39,7 @@ require_once __DIR__ . '/include/functions.render.php';
 // irmtfan use require_once because it will redeclared in newbb/blocks/list_topic.php
 //require_once __DIR__ . '/./class/TopicRenderer.php';
 $topicRenderer            = TopicRenderer::getInstance();
-$topicRenderer->userlevel = $GLOBALS['xoopsUserIsAdmin'] ? 2 : is_object($GLOBALS['xoopsUser']);
+$topicRenderer->userlevel = $GLOBALS['xoopsUserIsAdmin'] ? 2 : (int)is_object($GLOBALS['xoopsUser']);
 // irmtfan if list topic block is in the page then force to parse
 if (defined('LIST_TOPIC_DEFINED')) {
     $topicRenderer->force = true; // force against static vars
